refactor: centralize loops config parsing in utils module
